implementando autenticação na api no modulo categoria. somente usuário logado pode incluir registros no modulo categoria. se não estiver  logado não inclui o registro validação por cokkie e token criptografados para evitar interceptações.

+++ api/classes/categoria.php
@@ -1,5 +1,6 @@
 <?php
 require "classeBase.php";
+
 /**
 * classe Categoria
 *
@@ -12,9 +13,9 @@ class Categoria extends classeBase
     * @access private
     * @var integer Armazena o codigo da categoria
     */
-	private $codigoCategoria;
+    private $codigoCategoria;
 
-	/**
+    /**
     * @access private
     * @var string Armazena o nome da categoria
     */
@@ -23,7 +24,8 @@ class Categoria extends classeBase
     function __construct()
     {
         require_once '../adm/classes/conexao.php';
-    }    
+    }
+    
     /**
     * metodo acessor Get que retorna a informação da propriedade codigoCategoria
     *
@@ -34,12 +36,12 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */ 	
-	public function getCodigoCategoria() 
-	{
-		return $this->codigoCategoria;
-	}//public function getCodigoCategoria() 
+    public function getCodigoCategoria() 
+    {
+        return $this->codigoCategoria;
+    }//public function getCodigoCategoria() 
 
-	/**
+    /**
     * metodo acessor Set que carrega informação na propriedade codigoCategoria
     *
     * @access    public
@@ -49,12 +51,12 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
-	public function setCodigoCategoria($codigoCategoria) 
-	{
-		$this->codigoCategoria = $codigoCategoria;
-	}//public function setCodigoCategoria($codigoCategoria)
+    public function setCodigoCategoria($codigoCategoria) 
+    {
+        $this->codigoCategoria = $codigoCategoria;
+    }//public function setCodigoCategoria($codigoCategoria)
 
-	/**
+    /**
     * metodo acessor Get que retorna a informação da propriedade nome
     *
     * @access    public
@@ -64,12 +66,12 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
-	public function getNome() 
-	{
-		return $this->nome;
-	}//public function getNome()
+    public function getNome() 
+    {
+        return $this->nome;
+    }//public function getNome()
 	
-	/**
+    /**
     * metodo acessor Set que carrega informação na propriedade nome
     *
     * @access    public
@@ -79,12 +81,12 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
-	public function setNome($nome) 
-	{
-		$this->nome = $nome;
-	}//public function setNome($nome)
+    public function setNome($nome) 
+    {
+        $this->nome = $nome;
+    }//public function setNome($nome)
 
-	/**
+    /**
     * metodo que tem função de fazer gravação do registro
     *
     * @access    public
@@ -94,40 +96,40 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
-	public function inserir($token)
-	{
-                if ($this->tokenEhValido($token) === false) {
-                    return false;
-                }
-                        
-		try {
-			$nome  = $this->getNome();
-			$id    = $this->ultimoCodigo();
+    public function inserir($token)
+    {
+        try {
+            if ($this->tokenEhValido($token) === false) {
+                return false;
+            }//if ($this->tokenEhValido($token) === false) {
+        
+            $nome  = $this->getNome();
+            $id    = $this->ultimoCodigo();
 
-		  	$sql  = "\n INSERT INTO categoria (";
-			$sql .= "\n   `codigo_categoria`";
-			$sql .= "\n , `nome`";
-			$sql .= "\n ) VALUES (";
-			$sql .= "\n   :id";
-			$sql .= "\n , :nome";
-			$sql .= "\n )";
-			
-			$conexao = Conexao::getConexao();	
-			$conexao->beginTransaction();
-			$stmt = $conexao->prepare($sql);
-			$stmt->bindParam(":id",$id);
-			$stmt->bindParam(":nome",$nome);
-			$retorno = $stmt->execute();
-			$conexao->commit();
-			$conexao = null;
-			return $retorno;
-		} catch (PDOException $e) {
-			$conexao = null;
-			return $e->getMessage();
-		}
-	}//public function inserir()
+            $sql  = "\n INSERT INTO categoria (";
+            $sql .= "\n   `codigo_categoria`";
+            $sql .= "\n , `nome`";
+            $sql .= "\n ) VALUES (";
+            $sql .= "\n   :id";
+            $sql .= "\n , :nome";
+            $sql .= "\n )";
+
+            $conexao = Conexao::getConexao();	
+            $conexao->beginTransaction();
+            $stmt = $conexao->prepare($sql);
+            $stmt->bindParam(":id",$id);
+            $stmt->bindParam(":nome",$nome);
+            $retorno = $stmt->execute();
+            $conexao->commit();
+            $conexao = null;
+            return $retorno;
+        } catch (PDOException $e) {
+            $conexao = null;
+            return $e->getMessage();
+        }
+    }//public function inserir()
 	
-	/**
+    /**
     * metodo que tem função de fazer alteração do registro
     *
     * @access    public
@@ -137,32 +139,36 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
-    public function alterar()
-	{
-		try {
-			$codigo  = $this->getCodigoCategoria();
-			$nome    = $this->getNome();
+    public function alterar($token)
+    {
+        try {
+            if ($this->tokenEhValido($token) === false) {
+                return false;
+            }//if ($this->tokenEhValido($token) === false) {
 
-			$sql  = "\n UPDATE categoria"; 
-			$sql .= "\n SET nome = :nome";
-			$sql .= "\n WHERE codigo_categoria = :id";
+            $codigo  = $this->getCodigoCategoria();
+            $nome    = $this->getNome();
 
-			$conexao = Conexao::getConexao();
-			$conexao->beginTransaction();
-			$stmt = $conexao->prepare($sql);
-			$stmt->bindParam(":id",$codigo);
-			$stmt->bindParam(":nome",$nome);
-			$retorno = $stmt->execute();
-			$conexao->commit();
-			$conexao = null;
-			return $retorno;
-		} catch (PDOException $e) {
-		  $conexao = null;
-		  return $e->getMessage();
-		}
-	}//public function alterar($codigo)
+            $sql  = "\n UPDATE categoria"; 
+            $sql .= "\n SET nome = :nome";
+            $sql .= "\n WHERE codigo_categoria = :id";
+
+            $conexao = Conexao::getConexao();
+            $conexao->beginTransaction();
+            $stmt = $conexao->prepare($sql);
+            $stmt->bindParam(":id",$codigo);
+            $stmt->bindParam(":nome",$nome);
+            $retorno = $stmt->execute();
+            $conexao->commit();
+            $conexao = null;
+            return $retorno;
+        } catch (PDOException $e) {
+            $conexao = null;
+            return $e->getMessage();
+        }
+    }//public function alterar($codigo)
 	
-	/**
+    /**
     * metodo que tem função de fazer exclusão do registro
     *
     * @access    public
@@ -172,30 +178,34 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
-	public function excluir()
-	{
-		try {
-			$codigo  = $this->getCodigoCategoria();
-			$nome    = $this->getNome();
+    public function excluir($token)
+    {
+        try {
+            if ($this->tokenEhValido($token) === false) {
+                return false;
+            }//if ($this->tokenEhValido($token) === false) {
+            
+            $codigo  = $this->getCodigoCategoria();
+            $nome    = $this->getNome();
 
-			$sql  = "\n DELETE FROM categoria";
-			$sql .= "\n WHERE codigo_categoria = :id";
+            $sql  = "\n DELETE FROM categoria";
+            $sql .= "\n WHERE codigo_categoria = :id";
 
-			$conexao = Conexao::getConexao();
-			$conexao->beginTransaction();
-			$stmt = $conexao->prepare($sql);
-			$stmt->bindParam(":id",$codigo);
-			$retorno = $stmt->execute();
-			$conexao->commit();
-			$conexao = null;
-			return $retorno;
-		} catch (PDOException $e) {
-			$conexao = null;
-			return $e->getMessage();
-		}
-	}//public function excluir($codigo)
+            $conexao = Conexao::getConexao();
+            $conexao->beginTransaction();
+            $stmt = $conexao->prepare($sql);
+            $stmt->bindParam(":id",$codigo);
+            $retorno = $stmt->execute();
+            $conexao->commit();
+            $conexao = null;
+            return $retorno;
+        } catch (PDOException $e) {
+            $conexao = null;
+            return $e->getMessage();
+        }
+    }//public function excluir($codigo)
 	
-	/**
+    /**
     * metodo que tem função de buscar o ultimo codigo da categoria
     *
     * @access    public
@@ -205,26 +215,25 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
-	public static function ultimoCodigo()
-	{
-		try {
-                        
-			$conexao = Conexao::getConexao();
-			$sql  = "\n SELECT MAX(codigo_categoria) + 1 AS codigo";
-			$sql .= "\n FROM categoria";
+    public static function ultimoCodigo()
+    {
+        try {
+            $conexao = Conexao::getConexao();
+            $sql  = "\n SELECT MAX(codigo_categoria) + 1 AS codigo";
+            $sql .= "\n FROM categoria";
 
-			$stmt = $conexao->prepare($sql);
-			$stmt->execute();
-			$retorno =  $stmt->fetch(PDO::FETCH_ASSOC);
-			$conexao = null;
-			return $retorno["codigo"];
- 		} catch (PDOException $e) {
- 			$conexao = null;
-			return $e->getMessage();
-		}
-	}//public static function ultimoCodigo()
+            $stmt = $conexao->prepare($sql);
+            $stmt->execute();
+            $retorno =  $stmt->fetch(PDO::FETCH_ASSOC);
+            $conexao = null;
+            return $retorno["codigo"];
+        } catch (PDOException $e) {
+            $conexao = null;
+            return $e->getMessage();
+        }
+    }//public static function ultimoCodigo()
 
-	/**
+    /**
     * metodo que tem função de buscar as informacoes da categoria por codigo
     *
     * @access    public
@@ -234,29 +243,29 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
-	public static function listarPorCodigo($codigo)
-	{
-		try {
-                        require_once '../classes/conexao.php';
-			$conexao = Conexao::getConexao();
+    public static function listarPorCodigo($codigo)
+    {
+        try {
+            require_once '../adm/classes/conexao.php';
+            $conexao = Conexao::getConexao();
 
-			$sql  = "\n SELECT codigo_categoria";
-			$sql .= "\n , nome";
-			$sql .= "\n FROM categoria";
-			$sql .= "\n WHERE codigo_categoria = :id";
+            $sql  = "\n SELECT codigo_categoria";
+            $sql .= "\n , nome";
+            $sql .= "\n FROM categoria";
+            $sql .= "\n WHERE codigo_categoria = :id";
 
-			$stmt = $conexao->prepare($sql);
-			$stmt->bindParam(":id",$codigo);
-			$stmt->execute();
-			$conexao = null;
-			return $stmt->fetch(PDO::FETCH_ASSOC);
-		} catch (PDOException $e) {
-			$conexao = null;
-			return $e->getMessage();
-		}		
-	}//public static function listarPorCodigo($codigo)
+            $stmt = $conexao->prepare($sql);
+            $stmt->bindParam(":id",$codigo);
+            $stmt->execute();
+            $conexao = null;
+            return $stmt->fetch(PDO::FETCH_ASSOC);
+        } catch (PDOException $e) {
+            $conexao = null;
+            return $e->getMessage();
+        }		
+    }//public static function listarPorCodigo($codigo)
 	
-	/**
+    /**
     * metodo que tem função de buscar todas as categorias que tenham determinado padrao de nome
     *
     * @access    public
@@ -266,31 +275,31 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
- 	public static function listarPorNome($nome)
-	{
-		try {
-                        require_once '../classes/conexao.php';
-			$nome .= "%";
-			$conexao = Conexao::getConexao();
+    public static function listarPorNome($nome)
+    {
+        try {
+            require_once '../classes/conexao.php';
+            $nome .= "%";
+            $conexao = Conexao::getConexao();
 
-			$sql  = "\n SELECT codigo_categoria";
-			$sql .= "\n , nome";
-			$sql .= "\n FROM categoria";
-			$sql .= "\n WHERE nome LIKE :nome";
+            $sql  = "\n SELECT codigo_categoria";
+            $sql .= "\n , nome";
+            $sql .= "\n FROM categoria";
+            $sql .= "\n WHERE nome LIKE :nome";
 
 
-			$stmt = $conexao->prepare($sql);
-			$stmt->bindParam(":nome",$nome);
-			$stmt->execute();
-			$conexao = null;
-			return $stmt->fetchAll(PDO::FETCH_ASSOC);
-		} catch (PDOException $e) {
-			$conexao = null;
-			return $e->getMessage();
-		}	
-	}//public static function listarPorNome($nome)
+            $stmt = $conexao->prepare($sql);
+            $stmt->bindParam(":nome",$nome);
+            $stmt->execute();
+            $conexao = null;
+            return $stmt->fetchAll(PDO::FETCH_ASSOC);
+        } catch (PDOException $e) {
+            $conexao = null;
+            return $e->getMessage();
+        }	
+    }//public static function listarPorNome($nome)
 
-	/**
+    /**
     * metodo que tem função de buscar todas as categorias
     *
     * @access    public
@@ -300,25 +309,24 @@ class Categoria extends classeBase
     * @since     14/12/2010
     * @version   0.2
     */
-	public static function listarTudo($strRequire = '../adm/classes/conexao.php')
-	{
-                require_once $strRequire;
-                
-		try {
-                        
-			$conexao = Conexao::getConexao();
+    public static function listarTudo($strRequire = '../adm/classes/conexao.php')
+    {
+        require_once $strRequire;
 
-			$sql  = "\n SELECT codigo_categoria";
-			$sql .= "\n , nome";
-			$sql .= "\n FROM categoria";
+        try {
+            $conexao = Conexao::getConexao();
 
-			$stmt = $conexao->prepare($sql);
-			$stmt->execute();
-			$conexao = null;
-			return $stmt->fetchAll();
-		} catch (PDOException $e) {
-			$conexao = null;	
-			return $e->getMessage();
-		}	
-	}//public static function listarTudo()
+            $sql  = "\n SELECT codigo_categoria";
+            $sql .= "\n , nome";
+            $sql .= "\n FROM categoria";
+
+            $stmt = $conexao->prepare($sql);
+            $stmt->execute();
+            $conexao = null;
+            return $stmt->fetchAll();
+        } catch (PDOException $e) {
+            $conexao = null;	
+            return $e->getMessage();
+        }	
+    }//public static function listarTudo()
 }
